+++ apps/builddao/widget/components/modals/propose/RemoveMember.jsx
@@ -197,17 +197,6 @@ useEffect(() => {
   return () => clearTimeout(timeoutId);
 }, [copied]);
 
-const handleCopy = () => {
-  clipboard
-    .writeText(url)
-    .then(() => {
-      setCopied(true);
-    })
-    .catch((error) => {
-      console.error("Failed to copy:", error);
-    });
-};
-
 return (
   <div className="d-flex flex-column">
     <div className="form-group mb-3">
@@ -296,7 +285,8 @@ return (
       open={isInfoModalActive}
       setInfoModalActive={setInfoModalActive}
       copied={copied}
-      onCopyButtonClick={handleCopy}
+      setCopied={setCopied}
+      url={url}
     />
   </div>
 );
