+++ apps/builddao/widget/components/modals/propose/Text.jsx
@@ -183,17 +183,6 @@ useEffect(() => {
   return () => clearTimeout(timeoutId);
 }, [copied]);
 
-const handleCopy = () => {
-  clipboard
-    .writeText(url)
-    .then(() => {
-      setCopied(true);
-    })
-    .catch((error) => {
-      console.error("Failed to copy:", error);
-    });
-};
-
 return (
   <div className="d-flex flex-column ">
     <label>Proposal Description</label>
@@ -236,11 +225,12 @@ return (
         Create
       </Button>
     </div>
-     <ProposalVisibilityInfoModal
+    <ProposalVisibilityInfoModal
       open={isInfoModalActive}
       setInfoModalActive={setInfoModalActive}
       copied={copied}
-      onCopyButtonClick={handleCopy}
+      setCopied={setCopied}
+      url={url}
     />
   </div>
 );
